requirements.txt
flask>=2.3.0
python-docx>=0.8.11
Werkzeug>=2.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
pillow>=10.0.0 